+++ starsky/starsky.feature.thumbnail/Interfaces/IThumbnailGenerationService.cs
@@ -1,0 +1,8 @@
+using System.Threading.Tasks;
+
+namespace starsky.feature.thumbnail.Interfaces;
+
+public interface IThumbnailGenerationService
+{
+	Task BgQueue(string subPath);
+}
