+++ starsky/starsky.sln
@@ -148,6 +148,8 @@ Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "starskydemoseedcli", "stars
 EndProject
 Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "documentation", "..\documentation\documentation.csproj", "{B7CB2C57-117D-43B4-9884-59BEAC3B6DB2}"
 EndProject
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "starsky.feature.thumbnail", "starsky.feature.thumbnail\starsky.feature.thumbnail.csproj", "{5B70A0EC-1E7A-496E-94C7-A7ED1628B7EF}"
+EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
 		Debug|Any CPU = Debug|Any CPU
@@ -338,6 +340,10 @@ Global
 		{B7CB2C57-117D-43B4-9884-59BEAC3B6DB2}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{B7CB2C57-117D-43B4-9884-59BEAC3B6DB2}.Release|Any CPU.ActiveCfg = Release|Any CPU
 		{B7CB2C57-117D-43B4-9884-59BEAC3B6DB2}.Release|Any CPU.Build.0 = Release|Any CPU
+		{5B70A0EC-1E7A-496E-94C7-A7ED1628B7EF}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{5B70A0EC-1E7A-496E-94C7-A7ED1628B7EF}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{5B70A0EC-1E7A-496E-94C7-A7ED1628B7EF}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{5B70A0EC-1E7A-496E-94C7-A7ED1628B7EF}.Release|Any CPU.Build.0 = Release|Any CPU
 	EndGlobalSection
 	GlobalSection(SolutionProperties) = preSolution
 		HideSolutionNode = FALSE
@@ -392,5 +398,6 @@ Global
 		{63CF4BEC-DCC2-4688-8911-FB4B510032D6} = {4B9276C3-651E-48D3-B3A7-3F4D74F3D01A}
 		{C4219438-A401-47D5-9CF3-A198BB6E3185} = {B974FC20-C3EE-4EB0-AF25-F0D8DA2C28D7}
 		{B7CB2C57-117D-43B4-9884-59BEAC3B6DB2} = {2F68B3F8-5277-45CF-A820-50F93E7BF7D9}
+		{5B70A0EC-1E7A-496E-94C7-A7ED1628B7EF} = {4B9276C3-651E-48D3-B3A7-3F4D74F3D01A}
 	EndGlobalSection
 EndGlobal
