+++ starsky/starsky/starsky.csproj
@@ -136,6 +136,7 @@
         <ProjectReference Include="..\starsky.feature.rename\starsky.feature.rename.csproj" />
         <ProjectReference Include="..\starsky.feature.search\starsky.feature.search.csproj" />
         <ProjectReference Include="..\starsky.feature.syncbackground\starsky.feature.syncbackground.csproj" />
+        <ProjectReference Include="..\starsky.feature.thumbnail\starsky.feature.thumbnail.csproj" />
         <ProjectReference Include="..\starsky.feature.webhtmlpublish\starsky.feature.webhtmlpublish.csproj" />
         <ProjectReference Include="..\starsky.foundation.accountmanagement\starsky.foundation.accountmanagement.csproj" />
         <ProjectReference Include="..\starsky.foundation.database\starsky.foundation.database.csproj" />
